first grade

+++ Keepr.client/src/components/Vault.vue
@@ -1,43 +1,39 @@
 <template>
-  <router-link :to="{name: 'Vault', params: {id: vault.id}}" @click="setActiveVault">
-    <div class="Vault card m-3 col-sm-3 shadow" v-if="vault.isPrivate == false">
-      <div class="p-3">
-        <h3 class="text-primary text-center">
-          <u>{{ vault.name }}</u>
-          <u></u>
-        </h3>
-        <h4 class="text-center p-3">
-          {{ vault.description }}
-        </h4>
-        <div class="d-flex justify-content-center">
-          <button @click="deleteVault(vault.id)" v-if="vault.creatorId === account.id" class="btn btn-danger" title="Delete Vault">
-            Delete
-          </button>
+  <div v-if="vault.isPrivate == false" class="vault-bg">
+    <router-link :to="{name: 'Vault', params: {id: vault.id}}" @click="setActiveVault">
+      <div class="Vault card m-3 shadow">
+        <div class="p-3 my-3">
+          <h5 class="vault-name text-primary text-center">
+            <u>{{ vault.name }}</u>
+          </h5>
         </div>
       </div>
+    </router-link>
+    <div class="d-flex justify-content-center">
+      <button @click="deleteVault(vault.id)" v-if="vault.creatorId === account.id" class="btn btn-danger" title="Delete Vault">
+        Delete
+      </button>
     </div>
-  </router-link>
-  <router-link :to="{name: 'Vault', params: {id: vault.id}}" @click="setActiveVault">
-    <div class="Vault card m-3 shadow" v-if="vault.isPrivate == true">
-      <div class="p-3">
-        <h3 class="text-primary text-center">
-          <u>{{ vault.name }}</u>
-          <u></u>
-        </h3>
-        <h4 class="text-center p-3">
-          {{ vault.description }}
-        </h4>
-        <div>
-          <p>Private man!</p>
-        </div>
-        <div class="d-flex justify-content-center">
-          <button @click="deleteVault(vault.id)" v-if="vault.creatorId === account.id" class="btn btn-danger" title="Delete Vault">
-            Delete
-          </button>
+  </div>
+  <div v-if="vault.isPrivate == true" class="vault-bg">
+    <router-link :to="{name: 'Vault', params: {id: vault.id}}" @click="setActiveVault">
+      <div class="Vault card m-3 shadow">
+        <div class="p-3 my-3">
+          <h5 class="vault-name text-primary text-center">
+            <u>{{ vault.name }}</u>
+            <div>
+              <p>Private man!</p>
+            </div>
+          </h5>
         </div>
       </div>
+    </router-link>
+    <div class="d-flex justify-content-center">
+      <button @click="deleteVault(vault.id)" v-if="vault.creatorId === account.id" class="btn btn-danger" title="Delete Vault">
+        Delete
+      </button>
     </div>
-  </router-link>
+  </div>
 </template>
 
 <script>
@@ -55,7 +51,9 @@ export default {
       state,
       account: computed(() => AppState.account),
       async deleteVault(id) {
-        await vaultsService.deleteVault(id)
+        if (confirm('do you really want to delete')) {
+          await vaultsService.deleteVault(id)
+        }
       },
       setActiveVault() {
         vaultsService.setActiveVault(props.vault.id)
@@ -66,5 +64,18 @@ export default {
 </script>
 
 <style scoped>
+ .Vault {
+  background-image: url("../assets/img/vault.jpeg");
+  background-size: cover;
+}
+.vault-name {
+  background-color: rgba(0, 0, 0, 0.5);
+}
+
+.vault-bg{
+  transition: all .2s ease-in-out;
+}
+.vault-bg:hover {
+  transform: scale(1.1); }
 
 </style>
